data-service/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
kafka-python==2.0.2
Werkzeug>=2.0